Make OSDetectorModule external
